Little cleanup now that all tests are passing local and travis
