+++ scripts/e2e-servers.js
@@ -1,6 +1,58 @@
 #!/usr/bin/env node
-console.log(`
-Before running this script, don't forget to start webdriver:
+
+const fs = require('fs');
+const { fork } = require('child_process');
+
+const constants = require('./../tests/constants');
+const utils = require('../tests/utils');
+
+const WRITE_TIMESTAMPS = false;
+const WRITE_TO_CONSOLE = false;
+
+const writeToStream = (stream, data) => {
+  const formattedData = WRITE_TIMESTAMPS ? `${new Date().toISOString()} ${data}` : data.toString();
+  stream.write(formattedData);
+
+  if (WRITE_TO_CONSOLE) {
+    console.log(formattedData);
+  }
+};
+
+const startServer = serviceName => new Promise((resolve, reject) => {
+  if(!fs.existsSync('tests/logs')) {
+    fs.mkdirSync('tests/logs');
+  }
+  
+  try {
+    const logStream = fs.createWriteStream(`tests/logs/${serviceName}.e2e.log`, { flags:'w' });
+
+    const server = fork(`${serviceName}/server.js`, {
+      stdio: 'pipe',
+      detached: false,
+      env: {
+        TZ: 'UTC',
+        API_PORT: constants.API_PORT,
+        COUCH_URL: utils.getCouchUrl(),
+        COUCH_NODE_NAME: process.env.COUCH_NODE_NAME,
+        PATH: process.env.PATH,
+      },
+    });
+
+    const writeToLogStream = data => writeToStream(logStream, data);
+    server.stdout.on('data', writeToLogStream);
+    server.stderr.on('data', writeToLogStream);
+    server.on('close', code => writeToLogStream(`${serviceName} process exited with code ${code}`));
+    resolve();
+  } catch (err) {
+    reject(err);
+  }
+});
+
+const startApi = () => startServer('api');
+const startSentinel = () => startServer('sentinel');
+const startAll = () => Promise.all([startApi(), startSentinel()]);
+
+console.log(`Before running this script, don't forget to start webdriver:
 
 	npm run webdriver
 
@@ -11,5 +63,4 @@ To see log files:
 
 Starting e2e test services…`);
 
-const serviceManager = require('./service-manager');
-serviceManager.startAll().then(() => console.log('[e2e] All services started.'));
+startAll().then(() => console.log('[e2e] All services started.'));
